rename pkg utils -> str -> xstrings for extended strings package
