+++ xstrings/xstrings.go
@@ -1,4 +1,4 @@
-package str
+package xstrings
 
 //Substr extract string of length in s starting at pos
 func Substr(s string, pos, length int) string {
